pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_court"
version = "0.1.0"
description = "AI Court - Legal Case Analysis and Prediction"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "AI Court Team", email = "dev@aicourt.com"}
]
dependencies = [
    "Flask>=2.2",
    "flask-cors>=3.0",
    "dill>=0.3",
    "numpy>=1.21",
    "pydantic>=2.7",
    "pandas>=1.3",
    "pyarrow>=14.0",
    "scikit-learn>=1.2",
    "nltk>=3.8",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "transformers>=4.39",
    "sentence-transformers>=2.2",
    "torch>=2.1",
    "accelerate>=0.26.0",
    "python-dotenv>=1.0",
    "flasgger>=0.9.7.1",
    "Flask-Limiter>=3.8",
    "prometheus-client>=0.20",
    "sentry-sdk[flask]>=2.13",
    "pyyaml>=6.0",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "mlflow>=2.14"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.5",
    "mypy>=1.8"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
